+++ trunk/openvibe-plugins/signal-processing/trunc/src/box-algorithms/basic/ovpCBoxAlgorithmSignalDecimation.cpp
@@ -121,7 +121,7 @@ boolean CBoxAlgorithmSignalDecimation::process(void)
 			}
 
 			m_ui32OutputSampleIndex=0;
-			m_ui32OutputSampleCountPerSentBlock=(m_ui32InputSampleCountPerSentBlock/m_i64DecimationFactor);
+			m_ui32OutputSampleCountPerSentBlock=(uint32)(m_ui32InputSampleCountPerSentBlock/m_i64DecimationFactor);
 			m_ui32OutputSampleCountPerSentBlock=(m_ui32OutputSampleCountPerSentBlock?m_ui32OutputSampleCountPerSentBlock:1);
 			m_ui64OutputSamplingFrequency=op_ui64SamplingRate/m_i64DecimationFactor;
